// src/app/interfaces/car.ts
import { Fuel } from "../enums/fuel";

export interface Car {
    id: number;
    title: string;
    address: string;
    brand: string;
    model: string;
    model_year: number;
    issuance: string;
    mileage: number;
    fuel: Fuel;
    color: string;
    numbers_doors: number;
    horse_power: number;
    price: number;
    pictures: string[];
    sold: boolean;

}